rename(feature/#16): 파일명에 있는 공백 제거
